A good run with an oblique ridge
  w->random_seed=203540935;
  w->ridge_radius=0.200000;
  w->c2=2.000000; w->c3=0.450000;
  w->extra_mutation_rate=0.100000;
  w->crossover_freq=0.300000;
  w->edge_inheritance=INHERIT_HALF_OF_EDGES_FROM_BOTH_PARENTS;

  //w->num_organisms=40;
  w->num_candidates=7;
  w->generations_per_epoch=20;
  w->sa_timesteps=20;

These parameters hill-climb an oblique ridge at y=2x+0.45. Good
hill-climbing starts in epoch 8 and continues for only a few more
epochs. These parameters are saved in good_run_oblique().

General observation (not rigorously checked out): even when the
parameters are right to hill-climb an oblique ridge, it doesn't
usually happen all that often. When it does happen, it starts in
early epochs: sometimes as early as epoch 1, maybe as late as
epoch 35. In *most* runs, signs of hill-climbing seem to appear
shortly after the population finds the ridge. Then the population
soon gets stuck at a certain phenotype. After a couple hundred
epochs, usually the phenotype is extremely stable.