tom
A strange run: it crawls up the side of the phenotype fitness
function, resulting in a bumpless virtual fitness function.
But how does it track motion of the peak from epoch to epoch?
It only has 5 nodes and 35 edges. Is it using the edges as a sort
of knob to turn?