Experimented with tweaking various parameters
The main result: if the ridge includes or comes near one of the
"easy" spreading-activation attractors--that is, phenotypes that
have -1.0, 0.0, or +1.0 for at least one coordinate--then the
population is likely to get stuck there. These are the circumstances
that occasionally produce a genotype that climbs the ridge sideways:
holding x=0.0 and climbing by varying y. That produces a bumpless
fitness landscape, but it can't follow the peak as it moves from
epoch to epoch. It just clings to one central bump that persists
from epoch to epoch.