Finished big refactoring. Orthogonal parameters.
There are now separate node parameters for various steps in the SA
algorithm, which overlapped in previous versions:
   INPUT_ACC                how inputs accumulate during one timestep
   ACTIVATION_TYPE          how accumulated inputs map to activation
   OUTPUT_TYPE              how activation maps to output
   CONTROL_UPDATE           how the 'control' knob updates
   INITIAL_ACTIVATION_TYPE  whether a node has initial activation

The program now compiles on OSX 10.7.5.

I commented out the "#ifdef WITH_SWIG" around main() so that I
could test the new code.

The measure of phenotype-function acclivity is now broken. We
probably need to add a new parameter to run climb_hill() on the
null_organism in phenotype_acclivity(). This new parameter could
simply allow phenotype nodes to have an initial_activation, which
sa doesn't modify.