--reward_coverage
--reward_coverage adds points to fitness to reward how well the range
of the g->ph function covers the part of the phenotype function that
has fitness > 1.0 (i.e the part above the bumps).

--reward_coverage slows the program down by a factor of something
like 2,500, due to the huge number of extra spreading activations
that need to be performed.

The purpose of --reward_coverage is to gain some insight into why
the program so often evolves g->ph functions (i.e. graphs) with tiny
ranges.